# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: VerifiA API
  version: 1.0.0
  description: |
    API REST para el sistema VerifiA de tokens de presencia criptográficos efímeros.

    ## Flujo completo
    1. Portal → `POST /challenges` → obtiene nonce + QR
    2. App escanea QR → realiza FaceTec + App Attest + Passkey
    3. App → `POST /tokens/issue` → recibe JWT ES256
    4. Portal hace polling → `GET /tokens/verify/:nonce`
    5. Portal → `POST /tokens/validate/:nonce` → consume token → muestra ✅/❌

  contact:
    name: Carlos Iván Ibarra Pacheco
    email: A01646019@tec.mx
servers:
  - url: https://api.verifia.dev
    description: Producción (Railway)
  - url: http://localhost:3001
    description: Desarrollo local
...
paths:
...
  /api/v1/receipts/{id}:
    get:
      tags: [Receipts]
      summary: Consultar recibo por id (no destructivo)
      description: |
        Búsqueda por id opaco (cuid), sin prueba de posesión del JWT del recibo.
        Devuelve **solo** el subconjunto público: `valid`, `status`, `verified_at`,
        `badge_valid_from`, `badge_valid_until`. **Nunca** incluye `subject_name` ni
        `nonce` en este nivel. Si el llamador es el dueño autenticado (Bearer cuya
        cuenta coincide con `account_id`), añade `subject_name` y el bloque `identity`.
        Rate limit dedicado (más estricto que el global).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      security:
        - BearerAuth: []
        - {}
      responses:
        '200':
          description: Estado del recibo (subconjunto según autenticación)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptStatusResponse'
        '404':
          description: Recibo no encontrado
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/receipts/verify:
    post:
      tags: [Receipts]
      summary: Verificar recibo por JWT completo
      description: |
        Verifica firma ES256 + propósito + expiración del JWT del recibo, y confirma
        que la fila sigue existiendo. No muta estado. Como el llamador envía el JWT
        firmado completo (prueba de posesión), es seguro devolver `subject_name`. El
        bloque `identity` sigue restringido al dueño autenticado.

        Distingue **expirado** (la verificación sí ocurrió, solo venció la constancia
        de 30 días → `status: EXPIRED`) de **inválido** (firma mala / manipulado →
        `status: INVALID`).
      security:
        - BearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receipt_jwt]
              properties:
                receipt_jwt:
                  type: string
                  description: JWT del recibo (verifia://receipt?jwt=...)
      responses:
        '200':
          description: Resultado de verificación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptStatusResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key del verificador. Define el claim `aud` del JWT.
        Cada organización verificadora recibe una clave única.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token de sesión de cuenta (JWT ES256, TTL 7 días). Emitido por
        `/auth/login` y `/auth/set-password`.
  schemas:
...
    ReceiptStatusResponse:
      type: object
      properties:
        valid: { type: boolean }
        status:
          type: string
          enum: [VALID, EXPIRED, INVALID, NOT_FOUND]
        verified_at: { type: string, format: date-time }
        badge_valid_from: { type: string, format: date-time }
        badge_valid_until: { type: string, format: date-time }
        subject_name:
          type: string
          nullable: true
          description: Solo vía POST /verify o nivel dueño autenticado
        identity:
          allOf:
            - $ref: '#/components/schemas/UserIdentity'
          nullable: true
          description: Solo nivel dueño autenticado
    UserIdentity:
      type: object
      properties:
        full_name: { type: string }
        curp: { type: string, nullable: true }
        date_of_birth: { type: string, nullable: true }
        id_type: { type: string, enum: [INE, PASSPORT] }
        profile_photo: { type: string, description: base64 JPEG }
        id_front_photo: { type: string, description: base64 JPEG }
        id_back_photo: { type: string, nullable: true }
        facetec_match_level: { type: integer, nullable: true }
        liveness_snapshot: { type: string, nullable: true }
        liveness_match_score: { type: integer, nullable: true }
...
    ApiError:
      type: object
      properties:
        error: { type: string }
        code: { type: string }
  responses:
    ValidationError:
      description: Payload inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Invalid request body"
            code: "VALIDATION_ERROR"
    RateLimited:
      description: Demasiadas peticiones
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Too many requests, please try again later."